refactor(cloudevents-server): refactor pkg struct
