+++ cloudevents-server/Dockerfile
@@ -1,4 +1,4 @@
-FROM golang:1.22.4 AS builder
+FROM golang:1.23.1 AS builder
 COPY . /app
 RUN cd /app && go build -o server
