--- cloudevents-server/pkg/events/custom/types.go
@@ -1,5 +1,0 @@
-package custom
-
-const (
-	EventTypeTestCaseRunReport = "test-case-run-report"
-)
