+++ cloudevents-server/pkg/events/tekton/handler_pipelinerun.go
@@ -1,7 +1,6 @@
 package tekton
 
 import (
-	"net/http"
 	"strings"
 
 	cloudevents "github.com/cloudevents/sdk-go/v2"
@@ -38,7 +37,7 @@ func (h *pipelineRunHandler) SupportEventTypes() []string {
 func (h *pipelineRunHandler) Handle(event cloudevents.Event) cloudevents.Result {
 	data := new(tektoncloudevent.TektonCloudEventData)
 	if err := event.DataAs(&data); err != nil {
-		return cloudevents.NewHTTPResult(http.StatusBadRequest, err.Error())
+		return cloudevents.NewReceipt(false, "invalid data: %v", err)
 	}
 
 	switch tektoncloudevent.TektonEventType(event.Type()) {
